cleanup

+++ run.sh
@@ -2,22 +2,46 @@
 
 set -e
 
+if [[ -n $DEBUG ]]; then
+  set -x
+fi
+
+function info() {
+  echo -n "[${app:unknown-app}] $*"
+}
+function ok() {
+  echo "ok"
+}
 
 # signal-cli
 (
+  app="signal-cli"
   cd docker/signal-cli
-  signalCliVersion=$(grep ^ARG Dockerfile | grep "SIGNAL_CLI_VERSION" | cut -d= -f2 )
+  signalCliVersion=$(grep ^ARG Dockerfile | grep "SIGNAL_CLI_VERSION" | cut -d= -f2)
   docker build -t signal-cli .
 
+  info "testing ... "
+  log=$(mktemp)
+  if ! docker run --rm -ti --entrypoint signal-cli -w /tmp signal-cli -v >"$log"; then
+    echo "ERROR signal-cli failed" >&2
+    cat "$log"
+    exit 1
+  fi
+  ok
+
+  info "exit code testing ... "
   log=$(mktemp)
   if docker run --rm -ti --entrypoint signal-cli signal-cli -u +49123456 receive >"$log" 2>&1; then
     echo "ERROR signal-cli failed" >&2
     cat "$log"
     exit 1
   fi
+  ok
 )
 
+mvnd -Dmaven.repo.local=.m2_repo clean verify
 docker build -t attachment-mover-java -f docker/attachment-mover-java/Dockerfile .
+
 log=$(mktemp)
 if ! docker run --rm -ti --entrypoint java attachment-mover-java -jar quarkus-run.jar -h >"$log"; then
   echo "ERROR app failed" >&2
@@ -28,7 +52,7 @@ fi
 if [[ -n $RELEASE ]]; then
   timestamp=$(date "+%Y%m%d-%H%M%S")
 
-  signalCliVersion=$(docker history --no-trunc signal-cli | grep -v ^$ | grep SIGNAL_CLI_VERSION | cut -d= -f2 | cut -d\  -f1| xargs)
+  signalCliVersion=$(docker history --no-trunc signal-cli | grep -v ^$ | grep SIGNAL_CLI_VERSION | cut -d= -f2 | cut -d\  -f1 | xargs)
   docker tag signal-cli lkwg82/signal-attachment-mover:signal-cli-"$signalCliVersion"
   docker push lkwg82/signal-attachment-mover:signal-cli-"$signalCliVersion"
   docker tag signal-cli lkwg82/signal-attachment-mover:signal-cli
